+++ CHANGES.md
@@ -8,6 +8,7 @@
 ## Furniture
 - added Furniture Handler -> Callback when Furniture gets loaded, which allows us to add custum scripts to the prefabs
 - added Furniture Shop Restock Handler -> sell your furnitures!
+- added Furniture Place Type
 - fixed Furniture stacking issues
 
 # v0.1.0-alpha (July 4th 2024) - Preview